Introduce `scalafmt-worker` module to address non-trivial API changes
The previous Java Reflection based approach are to inergonomic and error-prone since creating a Scalafmt now involves more API use.

Now, we directly depend on scalafmt-dynamic from the worker module.

The previous `ScalafmtWorker` class is now a trait, but keeps the same signature.
